@@ -42,7 +42,7 @@
     "postcss-loader": "^2.0.6",
     "react-hot-loader": "^3.0.0-beta.6",
     "style-loader": "^0.18.2",
-    "webpack": "^3.5.5",
+    "webpack": "^3.10.0",
     "webpack-dev-server": "^2.7.1"
   }
 }
