@@ -1,6 +1,7 @@
 import React, {Component} from 'react';
 import Spygame from './spygame';
 import UI from './ui';
+import openSocket from 'socket.io-client';
 import {connect} from 'react-redux';
 import {setConn, gameInfo} from "../actions"
 
@@ -8,25 +9,37 @@ class GameContainer extends Component{
 
     constructor(props){
         super(props);
+
+        const gameSocket = openSocket('localhost:8001', {
+            reconnection: false
+        });
+
+        gameSocket.on('player_event', (event) => {
+            console.log('player_event from GameContainer: ', event);
+        });
+
         this.state = {
-            role: null
+            role: null,
+            gameSocket: gameSocket,
         }
     }
 
-    componentWillMount(){
-        const eventId = 'join game';
-        const gameCreatorId = 'dummy1234';
-        const playerIds = {
-            currentPlayer: this.props.socketConnection.id,
-            gameCreatorPlayer: gameCreatorId
-        };
-        const gameId = this.props.openGame.placeId;
-        this.props.socketConnection.emit('join_button_pressed', eventId, gameId, playerIds);
+    componentDidMount(){
+        // const eventId = 'join game';
+        // const gameCreatorId = 'dummy1234';
+        // const playerIds = {
+        //     currentPlayer: this.props.socketConnection.id,
+        //     gameCreatorPlayer: gameCreatorId
+        // };
+        // const gameId = this.props.openGame.placeId;
+        // this.props.socketConnection.emit('join_button_pressed', eventId, gameId, playerIds);
 
         const socket = this.props.socketConnection;
 
-        socket.on('gameEnd',()=>{
+        socket.on('gameEnd',(thisGameID)=>{
             console.log('received game end notification');
+
+            // socket.emit('deleteGame', (missionName));
 
             this.props.history.push('/lobby');
         });
@@ -40,18 +53,23 @@ class GameContainer extends Component{
         });
     }
 
-    // componentWillUnmount(){
-    //     this.props.socketConnection.close();
-    // }
+    componentWillUnmount(){
+        // this.props.socketConnection.close();
+        const socket = this.props.socketConnection;
+        socket.removeListener('role');
+
+//         let uuid = socket.on("event", someHandler)
+// socket.off(id: uuid)
+    }
 
     render(){
         const role = this.state.role;
         return(
             <div>
                 <div id="gameContainer"  style={{pointerEvents: 'auto'}}>
-                    <Spygame />
+                    <Spygame gameSocket={this.state.gameSocket}/>
                 </div>
-                <UI role={role} style={{pointerEvents: 'none'}}/>
+                <UI role={role} gameSocket={this.state.gameSocket} style={{pointerEvents: 'none'}}/>
             </div>
         )
     }
@@ -62,7 +80,6 @@ function mapStateToProps(state){
         socketConnection: state.socketConnection.setConn,
         openGame: state.gameInformation.gameObject,
     }
-
 }
 
 export default connect(mapStateToProps, {setConn,gameInfo})(GameContainer)
