@@ -14,52 +14,78 @@ class spymasterUI extends Component {
     constructor(props){
         super(props);
 
-        const socket = this.props.socketConnection;
-
         this.state = {
-            page: 'Phase1'
+            page: 'Phase1',
+            event: '',
+            icon: '',
+            iconEventArray: [
+                {
+                    icon: 'camera',
+                    event: 'Camera detected agent'
+                },
+                // {
+                //     icon: 'https',
+                //     event: 'Door is locked'
+                // },
+                {
+                    icon: 'vpn_key',
+                    event: 'Door is unlocked'
+                },
+                {
+                    icon: 'radio_button_checked',
+                    event: 'Agent pressed switch'
+                },
+                {
+                    icon: 'pan_tool',
+                    event: 'Agent picked up item'
+                },
+                {
+                    icon: 'check_box',
+                    event: 'Mission Complete'
+                },
+                {
+                    icon: 'remove_red_eye',
+                    event: 'Guard detected agent'
+                },
+            ]
         };
+
+        console.log('gameSocket', this.props.gameSocket);
 
         this.setHtmlPage = this.setHtmlPage.bind(this);
         this.getHtmlPage = this.getHtmlPage.bind(this);
+    }
 
-        socket.on('player_event', (event) => {
+    componentDidMount() {
+        const gameSocket = this.props.gameSocket;
+
+        gameSocket.on('player_event', (event) => {
             console.log("player event", event);
-            switch(event){
-                case 'Camera detected agent':
-                    this.props.playerEvent('Camera detected agent', 'camera');
+
+            let iconName = '';
+
+            switch(event) {
+                case 'camera':
+                    iconName = 'camera'
                     break;
-                case 'Door is locked':
-                    this.props.playerEvent('Door is locked', 'vpn_key');
+                case 'door':
+                    iconName = 'vpn_key'
                     break;
-                case 'Door is unlocked':
-                    this.props.playerEvent('Door is unlocked', 'vpn_key');
+                case 'button':
+                    iconName = 'radio_button_checked'
                     break;
-                case 'Agent pressed switch':
-                    this.props.playerEvent('Agent pressed switch', 'radio_button_checked');
-                    break;
-                case 'Agent picked up item':
-                    this.props.playerEvent('Agent picked up item', 'pan_tool');
-                    break;
-                case 'Mission Complete':
-                    this.props.playerEvent('Mission Complete', 'check_box');
-                    break;
-                case 'Guard detected agent':
-                    this.props.player('Guard detected agent', 'remove_red_eye');
-                    break;
+                default:
+                    return null
             }
 
-            document.getElementById("spymaster_message").classList.add("spymaster_message_display_in");
-            document.getElementById("spymaster_message").setAttribute('style', 'display: inline-block');
+            let thisIcon = this.state.iconEventArray.find((iconTriggered) => {
+                return iconTriggered.icon === iconName;
+            });
 
-            setTimeout(() => {
-                document.getElementById("spymaster_message").classList.remove("spymaster_message_display_in");
-                document.getElementById("spymaster_message").classList.add("spymaster_message_display_out");
-                document.getElementById("spymaster_message").setAttribute('style', 'display: none');
-            }, 3000);
+            console.log("this icon", thisIcon);
 
+            this.setState({event: thisIcon.event, icon: thisIcon.icon})
         });
-
     }
 
     setHtmlPage(html){
@@ -100,9 +126,6 @@ class spymasterUI extends Component {
         return (
             <div id="spymasterUiContainer" style={{pointerEvents:'none'}}>
 
-                {/*Check to see if props are being passed correctly here*/}
-                {/*<Player />*/}
-                {/*<img id="monitor" src={monitor}/>*/}
                 <div id="spymasterFrame"> </div>
 
                 <div id="static_html_container">
@@ -122,40 +145,41 @@ class spymasterUI extends Component {
                     </div>
                 </div>
 
-                <div className="spymaster_icon_container" style={{top: '10vh'}}>
-                    <i className="material-icons spymaster_icons">camera</i>
-                    <p className="spymaster_event_text">Camera Detection</p>
-                </div>
+                {/*<div className="spymaster_icon_container" style={{top: '10vh'}}>*/}
+                    {/*<i className="material-icons spymaster_icons">camera</i>*/}
+                    {/*<p className="spymaster_event_text">Camera Detection</p>*/}
+                {/*</div>*/}
 
-                <div className="spymaster_icon_container" style={{top: '20vh'}}>
-                    <i className="material-icons spymaster_icons">vpn_key</i>
-                    <p className="spymaster_event_text">Door Locked/Door Unlocked</p>
-                </div>
+                {/*<div className="spymaster_icon_container" style={{top: '20vh'}}>*/}
+                    {/*<i className="material-icons spymaster_icons">vpn_key</i>*/}
+                    {/*<p className="spymaster_event_text">Door Locked/Door Unlocked</p>*/}
+                {/*</div>*/}
 
-                <div className="spymaster_icon_container" style={{top: '30vh'}}>
-                    <i className="material-icons spymaster_icons">radio_button_checked</i>
-                    <p className="spymaster_event_text">Switch Pressed</p>
-                </div>
+                {/*<div className="spymaster_icon_container" style={{top: '30vh'}}>*/}
+                    {/*<i className="material-icons spymaster_icons">radio_button_checked</i>*/}
+                    {/*<p className="spymaster_event_text">Switch Pressed</p>*/}
+                {/*</div>*/}
 
-                <div className="spymaster_icon_container" style={{top: '40vh'}}>
-                    <i className="material-icons spymaster_icons">pan_tool</i>
-                    <p className="spymaster_event_text">Picked up item</p>
-                </div>
+                {/*<div className="spymaster_icon_container" style={{top: '40vh'}}>*/}
+                    {/*<i className="material-icons spymaster_icons">pan_tool</i>*/}
+                    {/*<p className="spymaster_event_text">Picked up item</p>*/}
+                {/*</div>*/}
 
-                <div className="spymaster_icon_container" style={{top: '50vh'}}>
-                    <i className="material-icons spymaster_icons">check_box</i>
-                    <p className="spymaster_event_text">Completed Mission</p>
-                </div>
+                {/*<div className="spymaster_icon_container" style={{top: '50vh'}}>*/}
+                    {/*<i className="material-icons spymaster_icons">check_box</i>*/}
+                    {/*<p className="spymaster_event_text">Completed Mission</p>*/}
+                {/*</div>*/}
 
-                <div className="spymaster_icon_container" style={{top: '60vh'}}>
-                    <i className="material-icons spymaster_icons">remove_red_eye</i>
-                    <p className="spymaster_event_text">Guard detection</p>
-                </div>
+                {/*<div className="spymaster_icon_container" style={{top: '60vh'}}>*/}
+                    {/*<i className="material-icons spymaster_icons">remove_red_eye</i>*/}
+                    {/*<p className="spymaster_event_text">Guard detection</p>*/}
+                {/*</div>*/}
 
                 <div id="spymaster_message" style={{top: '70vh'}} >
 
-                    <i className="material-icons" id="spymaster_message_icon"> {this.props.event.icon}</i>
-                    <p id="spymaster_message_text">{this.props.event.event}</p>
+                    <i className="material-icons" id="spymaster_message_icon"> { this.state.icon } </i>
+                    <p id="spymaster_message_text"> { this.state.event } </p>
+
                 </div>
 
                 {/*<ComPanel id="leftPanel" />*/}
