@@ -1,6 +1,13 @@
 console.log("cp started");
 
+var express = require('express');
+var app = express();
+var http = require('http').Server(app);
+var io = require('socket.io')(http);
+const port = 8001;
+
 var gameObject = require('./helper/gameObject');
+gameObject.init(io);
 const get = require("./helper/calcFunctions");
 const db = require("./be/getMapCode");
 const mapTileDict = require('./helper/mapTileDict');
@@ -227,21 +234,17 @@ function harryInitMap() {
     harrySetCharStartPos(playerTracker[socketHolder2.id]);
 }
 
+// Variables for server functionality defined here. Will be organized later.
 
-
-
-var express = require('express');
-var app = express();
-var http = require('http').Server(app);
-var io = require('socket.io')(http);
-const port = 8001;
+// var express = require('express');
+// var app = express();
+// var http = require('http').Server(app);
+// var io = require('socket.io')(http);
+// const port = 8001;
 
 var eventMessage = "";
 
 // *****Global variables listed below should be transferred to Simulation object at a later time*****
-var randColor = ['blue', 'yellow', 'red', 'green', 'grey', 'purple'];
-var randColor1 = ['blue', 'red', 'green'];
-var randColor2 = ['yellow', 'black', 'purple'];
 
 var nameAdj = ['magnificent', 'vicious', 'friendly', 'cheerful', 'sad', 'happy', 'confused', 'lazy', 'jolly', 'effervescent', 'noble', 'cowardly', 'silly', 'thunderous', 'insightful', 'foolish', 'panicked', 'determined', 'awesome', 'sleepy', 'energetic', 'joyful', 'superior', 'alpha', 'courageous', 'far-sighted', 'limping', 'bumbling', 'serious', 'playful', 'cantankerous', 'stubborn', 'relaxed', 'laughing', 'coughing', 'blind', 'sublime', 'naked', 'ascended', 'swift', 'supreme', 'mad', 'silver', 'crimson', 'golden', 'silent', 'brash', 'crying'];
 var nameAnimal = ['octopus', 'tiger', 'chihuahua', 'shark', 'whale', 'hawk', 'eagle', 'leopard', 'cheetah', 'elephant', 'horse', 'beagle', 'piranha', 'platypus', 'ostrich', 'kakapo', 'parrot', 'wolf', 'snake', 'lizard', 'butterfly', 'frog', 'chameleon', 'fox', 'coyote', 'hummingbird', 'buffalo', 'chicken', 'hyena', 'lion', 'llama', 'alpaca', 'dove', 'mantis', 'owl', 'ox', 'squid', 'bat', 'capybara', 'bison', 'mammoth', 'chimp', 'hornet', 'squirrel', 'hamster', 'tortoise', 'raven', 'crow', 'dragon', 'unicorn', 'antelope', 'gazelle', 'giraffe', 'mongoose', 'weasel', 'badger'];
@@ -299,11 +302,16 @@ io.on('connection', function(socket){
         }
 
         if(playerTracker.length > 1){
+
+            console.log('>>>>>>>>>>> more than 1 player, grabbing map data! <<<<<<<<<<<');
+
             retrieveMapData();
             io.to('spymaster').emit( 'playerRole', 'spymaster');
             io.to('spy').emit('playerRole', 'spy');
         }
     });
+
+
 
     // console.log(playerTracker);
     // if(playerTracker.length === 1){
@@ -461,10 +469,12 @@ function PlayerObject(number, id, name, color, profilePic){
 
 function startSim(){
     console.log("Simulation has started!");
-    initializeMap();
 
-    simulationReference = setInterval(simulation, pollRate);
+    if(simulationReference === null){
+        initializeMap();
 
+        simulationReference = setInterval(simulation, pollRate);
+    }
 }
 
 function endProcess(){
@@ -477,6 +487,8 @@ function endSim(){
 
     clearInterval(simulationReference);
     console.log("Simulation has ended!");
+
+    simulationReference = null;
 }
 
 function simulation(){
@@ -725,6 +737,8 @@ function simUpdate(objToUpdate) {
                 endSim();
 
                 setTimeout(()=>{
+                    console.log('executing game restart');
+
                     playerTracker[socketHolder2.id].status.clickHistory = [];
                     playerTracker[socketHolder2.id].status.posX = charStartPos[0];
                     playerTracker[socketHolder2.id].status.posY = charStartPos[1];
@@ -748,22 +762,24 @@ function simUpdate(objToUpdate) {
             nextObject.update();
 
             if( checkCollide(objToUpdate, oldCoord, null, nextObject) ){
-                //Rebecca added for spymaster UI
-                io.to('spymaster').emit('player_event', 'Camera detected agent');
-                console.log('Camera detected agent');
 
-                finalSimState[3].set('MISSION FAILED! Restarting...');
-                nextObject.trigger(true);
+                // // io.emit('player_event', 'Camera detected agent');
+                // console.log('Camera detected agent');
 
-                endSim();
+                // finalSimState[3].set('MISSION FAILED! Restarting...');
+                // nextObject.trigger(true);
+                // nextObject.emit('spymaster');
+                // io.to('spymaster').emit('camera');
 
-                setTimeout(()=>{
-                    playerTracker[socketHolder2.id].status.clickHistory = [];
-                    playerTracker[socketHolder2.id].status.posX = charStartPos[0];
-                    playerTracker[socketHolder2.id].status.posY = charStartPos[1];
+                // endSim();
 
-                    startSim();
-                }, 3000)
+                // setTimeout(()=>{
+                //     playerTracker[socketHolder2.id].status.clickHistory = [];
+                //     playerTracker[socketHolder2.id].status.posX = charStartPos[0];
+                //     playerTracker[socketHolder2.id].status.posY = charStartPos[1];
+
+                //     startSim();
+                // }, 3000)
             }
 
             handlerSimState[2].push(nextObject);
@@ -894,6 +910,8 @@ function simUpdate(objToUpdate) {
                             if(nextObject.solid){
                                 if(nextObject.type === 'door'){
                                     if(nextObject.lockState === false){
+                                        io.to('spymaster').emit('door');
+                                        console.log('emit door')
                                         nextObject.animate = true;
                                         nextObject.solid = false;
                                     }
@@ -906,6 +924,7 @@ function simUpdate(objToUpdate) {
                                 // Currently a button named 'treasure' is the exit trigger
                                 // Must define a treasure gameObject later
                                 if(nextObject.name !== 'treasure'){
+                                    io.to('spymaster').emit('button');
                                     nextObject.trigger(false);
                                 }else{
                                     nextObject.display = false;
@@ -1068,7 +1087,8 @@ function checkCollide(objToUpdate, oldCoord, nextCoord, comparedObject ){
 
         if(solid && collide){
             console.log('circle collided!');
-            io.to('spymaster').emit('player_event', 'Guard detected agent');
+            // io.to('spymaster').emit('player_event', 'Guard detected agent');
+            io.to('spymaster').emit('guard');
             console.log('Guard detected agent');
             objToUpdate.status.clickHistory.push({x: objToUpdate.status.posX, y: objToUpdate.status.posY});
             return true;
